requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-genai>=1.0.0
cachetools>=5.3.0
python-dotenv>=1.0.0
